Resolve possible GSOD at Movielist when executed for the first time Traceback (most recent call last):   File '/usr/lib/enigma2/python/Components/ActionMap.py', line 56, in action     res = self.actions[action]()   File '/usr/lib/enigma2/python/Screens/MovieSelection.py', line 793, in btn_blue     self._callButton(config.movielist.btn_blue.value)   File '/usr/lib/enigma2/python/Components/config.py', line 2070, in __getattr__     raise AttributeError(name) AttributeError: btn_blue [ePyObject] (CallObject(<bound method ActionMap.action of <Components.ActionMap.HelpableActionMap object at 0xafe53940>>,('ColorActions', 'blue')) failed)
When executed for the first time initUserDefinedActions was not executed
yet als long we see the 'Waiting' message. When a user defimed button is
pressed in between then we could have a GSOD. By initializing the user
definitions in __init__ they are alway sinitialized. In addition I added
as extra a check not doing the user defined functions when the 'waiting'
text is visible.

Co-Authored-By: Littlesat <littlesat@openpli.org>